server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
openai>=1.30.0
pydantic>=2.6.0
python-dotenv>=1.0.1
httpx>=0.27.0
openai-agents==0.2.5
eval_type_backport>=0.2.0
